+++ js/components/generate/__tests__/generate.test.js
@@ -1,6 +1,7 @@
 import React from "react"
 import GenerateForm from "../GenerateForm.js"
-import { screen, render, fireEvent, waitFor, within } from "@testing-library/react"
+import { screen, render, waitFor, within } from "@testing-library/react"
+import { userEvent } from "@testing-library/user-event"
 
 describe("Handle invalid course inputs appropriately", () => {
   beforeEach(() => {
@@ -30,11 +31,18 @@ describe("Handle invalid course inputs appropriately", () => {
       expectedWarning: "Cannot generate graph – no courses entered!",
     },
   ])(".$coursesInputText", async ({ coursesInputText, expectedWarning }) => {
+    const user = userEvent.setup()
     const coursesInputField = screen.getByPlaceholderText("e.g., CSC207H1, CSC324H1")
-    fireEvent.change(coursesInputField, { target: { value: coursesInputText } })
+    await user.click(coursesInputField)
+    await user.tripleClick(coursesInputField)
+    if (coursesInputText === "") {
+      coursesInputText = "{Backspace}"
+    }
+    await user.keyboard(coursesInputText)
+
     expect(screen.queryByText("Invalid Course Input")).toBeNull()
     const genButton = screen.getByText("Generate Graph")
-    fireEvent.click(genButton)
+    await user.click(genButton)
     const warningModal = (await screen.findByText("Invalid Course Input")).parentElement
     expect(warningModal).not.toBeNull()
     const warningMessage = within(warningModal).queryByText(expectedWarning)
@@ -44,14 +52,18 @@ describe("Handle invalid course inputs appropriately", () => {
 })
 
 describe("Exiting the Warning Modal works", () => {
+  let user
   beforeEach(async () => {
+    user = userEvent.setup()
     render(<GenerateForm />)
     const coursesInputText = "MAT777H1"
     const coursesInputField = screen.getByPlaceholderText("e.g., CSC207H1, CSC324H1")
-    fireEvent.change(coursesInputField, { target: { value: coursesInputText } })
+    await user.click(coursesInputField)
+    await user.tripleClick(coursesInputField)
+    await user.keyboard(coursesInputText)
     expect(screen.queryByText("Invalid Course Input")).toBeNull()
     const genButton = screen.getByText("Generate Graph")
-    fireEvent.click(genButton)
+    await user.click(genButton)
     let warningModal = (await screen.findByText("Invalid Course Input")).parentElement
     expect(warningModal).not.toBeNull()
   })
@@ -59,7 +71,7 @@ describe("Exiting the Warning Modal works", () => {
   it("`Okay` button works", async () => {
     const okayButton = screen.getByText("Okay")
     expect(okayButton).toBeDefined()
-    fireEvent.click(okayButton)
+    await user.click(okayButton)
     await waitFor(() => {
       expect(screen.queryByText("Invalid Course Input")).toBeNull()
     })
@@ -67,7 +79,7 @@ describe("Exiting the Warning Modal works", () => {
   it("`X` button works", async () => {
     const XButton = screen.getByText("X")
     expect(XButton).toBeDefined()
-    fireEvent.click(XButton)
+    await user.click(XButton)
     await waitFor(() => {
       expect(screen.queryByText("Invalid Course Input")).toBeNull()
     })
@@ -75,32 +87,40 @@ describe("Exiting the Warning Modal works", () => {
 })
 
 it("No warning for valid course input strings", async () => {
+  const user = userEvent.setup()
   render(<GenerateForm />)
   const coursesInputText = "CSC443H1"
   const coursesInputField = screen.getByPlaceholderText("e.g., CSC207H1, CSC324H1")
-  fireEvent.change(coursesInputField, { target: { value: coursesInputText } })
+  await user.click(coursesInputField)
+  await user.tripleClick(coursesInputField)
+  await user.keyboard(coursesInputText)
   expect(screen.queryByText("Invalid Course Input")).toBeNull()
   const genButton = screen.getByText("Generate Graph")
-  fireEvent.click(genButton)
+  await user.click(genButton)
   await expect(screen.findByText("Invalid Course Input")).rejects.toThrow()
 })
 
 it("Submitting with valid courses and then making them invalid correctly updates Graph", async () => {
+  const user = userEvent.setup()
   render(<GenerateForm />)
   const coursesInputText = "CSC443H1"
   const coursesInputField = screen.getByPlaceholderText("e.g., CSC207H1, CSC324H1")
-  fireEvent.change(coursesInputField, { target: { value: coursesInputText } })
+  await user.click(coursesInputField)
+  await user.tripleClick(coursesInputField)
+  await user.keyboard(coursesInputText)
   expect(screen.queryByText("Invalid Course Input")).toBeNull()
   let genButton = screen.getByText("Generate Graph")
-  fireEvent.click(genButton)
+  await user.click(genButton)
   await expect(screen.findByText("Invalid Course Input")).rejects.toThrow()
   expect(screen.getByText("CSC443H1")).toBeDefined()
 
   // now submitting a bad input to see if the graph gets updated
   const coursesInputTextBad = "CSC443H7"
-  fireEvent.change(coursesInputField, { target: { value: coursesInputTextBad } })
+  await user.click(coursesInputField)
+  await user.tripleClick(coursesInputField)
+  await user.keyboard(coursesInputTextBad)
   expect(screen.queryByText("Invalid Course Input")).toBeNull()
-  fireEvent.click(genButton)
+  await user.click(genButton)
   const warningModal = await screen.findByText("Invalid Course Input")
   expect(warningModal).toBeDefined()
   expect(screen.queryByText("CSC443H1")).toBeNull()