+++ js/components/graph/__tests__/Edge.test.js
@@ -1,7 +1,8 @@
 import React from "react"
 import TestGraph from "./TestGraph"
-import { fireEvent, render, act, screen } from "@testing-library/react"
+import { render, act, screen } from "@testing-library/react"
 import Edge from "../Edge"
+import { userEvent } from "@testing-library/user-event"
 
 describe("Edge", () => {
   it("should match snapshot", async () => {
@@ -31,97 +32,105 @@ describe("Edge", () => {
       expect(path_h101_201.classList.contains("inactive")).toBe(true)
     })
     it("with selected source and unselected destination => 'takeable' Edge", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa101 = graph.getByTestId("aaa101")
       const path_101_201 = graph.getByTestId("h101->aaa201")
 
-      fireEvent.click(aaa101)
+      await user.click(aaa101)
       expect(path_101_201.classList.contains("takeable")).toBe(true)
     })
 
     it("with unselected source with selected destination => 'inactive' Edge", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa201 = graph.getByTestId("aaa201")
       const path_101_201 = graph.getByTestId("h101->aaa201")
       expect(path_101_201.classList.contains("inactive")).toBe(true)
 
-      fireEvent.click(aaa201)
+      await user.click(aaa201)
       expect(path_101_201.classList.contains("inactive")).toBe(true)
     })
 
     it("with selected source and destination => 'active' Edge", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa101 = graph.getByTestId("aaa101")
       const aaa201 = graph.getByTestId("aaa201")
       const path_101_201 = graph.getByTestId("h101->aaa201")
-      fireEvent.click(aaa101)
-      fireEvent.click(aaa201)
+      await user.click(aaa101)
+      await user.click(aaa201)
 
       expect(path_101_201.classList.contains("active")).toBe(true)
     })
 
     it("clicking a node changes the status of all out edges", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa102 = graph.getByTestId("aaa102")
       const path_102_bool1 = graph.getByTestId("aaa102->bool1")
       const path_102_bool2 = graph.getByTestId("aaa102->bool2")
       const path_bool2_aaa202 = graph.getByTestId("bool2->aaa202")
-      fireEvent.click(aaa102)
+      await user.click(aaa102)
       expect(path_102_bool1.classList.contains("takeable")).toBe(true)
       expect(path_102_bool2.classList.contains("active")).toBe(true)
       expect(path_bool2_aaa202.classList.contains("takeable")).toBe(true)
     })
 
     it("clicking a node and clicking it again resets the status of all edges", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa102 = graph.getByTestId("aaa102")
       const path_102_bool1 = graph.getByTestId("aaa102->bool1")
       const path_102_bool2 = graph.getByTestId("aaa102->bool2")
       const path_bool2_aaa202 = graph.getByTestId("bool2->aaa202")
-      fireEvent.click(aaa102)
+      await user.click(aaa102)
       expect(path_102_bool1.classList.contains("takeable")).toBe(true)
       expect(path_102_bool2.classList.contains("active")).toBe(true)
       expect(path_bool2_aaa202.classList.contains("takeable")).toBe(true)
-      fireEvent.click(aaa102)
+      await user.click(aaa102)
       expect(path_102_bool1.classList.contains("inactive")).toBe(true)
       expect(path_102_bool2.classList.contains("inactive")).toBe(true)
       expect(path_bool2_aaa202.classList.contains("inactive")).toBe(true)
     })
 
     it("clicking a node with a takeable edge makes that edge active", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa102 = graph.getByTestId("aaa102")
       const aaa202 = graph.getByTestId("aaa202")
       const path_bool2_aaa202 = graph.getByTestId("bool2->aaa202")
-      fireEvent.click(aaa102)
+      await user.click(aaa102)
       expect(path_bool2_aaa202.classList.contains("takeable")).toBe(true)
-      fireEvent.mouseEnter(aaa202)
+      await user.hover(aaa202)
       expect(path_bool2_aaa202.classList.contains("takeable")).toBe(true)
-      fireEvent.click(aaa202)
+      await user.click(aaa202)
       expect(path_bool2_aaa202.classList.contains("active")).toBe(true)
     })
 
     it("clicking reset, sets the selected edge with status takeable, inactive", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa101 = graph.getByTestId("aaa101")
       const path_101_201 = graph.getByTestId("h101->aaa201")
-      fireEvent.click(aaa101)
+      await user.click(aaa101)
       expect(path_101_201.classList.contains("takeable")).toBe(true)
 
-      fireEvent.click(graph.getByTestId("test-reset"))
+      await user.click(graph.getByTestId("test-reset"))
       expect(path_101_201.classList.contains("inactive")).toBe(true)
     })
 
     it("clicking reset, sets the selected edge with status active, inactive", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa101 = graph.getByTestId("aaa101")
       const aaa201 = graph.getByTestId("aaa201")
       const path_101_201 = graph.getByTestId("h101->aaa201")
-      fireEvent.click(aaa101)
-      fireEvent.click(aaa201)
+      await user.click(aaa101)
+      await user.click(aaa201)
       expect(path_101_201.classList.contains("active")).toBe(true)
 
-      fireEvent.click(graph.getByTestId("test-reset"))
+      await user.click(graph.getByTestId("test-reset"))
       expect(path_101_201.classList.contains("inactive")).toBe(true)
     })
   })
@@ -129,59 +138,65 @@ describe("Edge", () => {
   describe("hovering behaviour", () => {
     describe("hovering over source does nothing", () => {
       it("hovered and unselected source and unselected destination => stays 'inactive'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa101 = graph.getByTestId("aaa101")
         const path_101_201 = graph.getByTestId("h101->aaa201")
 
         expect(path_101_201.classList.contains("inactive")).toBe(true)
-        fireEvent.mouseOver(aaa101)
+        await user.hover(aaa101)
         expect(path_101_201.classList.contains("inactive")).toBe(true)
       })
 
       it("hovered and selected source and unselected destination => stays 'takeable'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa101 = graph.getByTestId("aaa101")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa101)
+        await user.click(aaa101)
         expect(path_101_201.classList.contains("takeable")).toBe(true)
-        fireEvent.mouseOver(aaa101)
+        await user.hover(aaa101)
         expect(path_101_201.classList.contains("takeable")).toBe(true)
       })
 
       it("hovered and unselected source and selected destination => stays 'active'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa101 = graph.getByTestId("aaa101")
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa201)
+        await user.click(aaa201)
         expect(path_101_201.classList.contains("inactive")).toBe(true)
-        fireEvent.mouseOver(aaa101)
+        await user.hover(aaa101)
         expect(path_101_201.classList.contains("inactive")).toBe(true)
       })
 
       it("(hovered and selected source) and selected destination => stays 'active'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa101 = graph.getByTestId("aaa101")
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa101)
-        fireEvent.click(aaa201)
+        await user.click(aaa101)
+        await user.click(aaa201)
         expect(path_101_201.classList.contains("active")).toBe(true)
-        fireEvent.mouseOver(aaa101)
+        await user.hover(aaa101)
         expect(path_101_201.classList.contains("active")).toBe(true)
       })
 
       it("hover on and hover off should change edge state to missing", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.mouseEnter(aaa201)
+        await user.hover(aaa201)
         expect(path_101_201.classList.contains("missing")).toBe(true)
-        fireEvent.mouseLeave(aaa201)
+        await user.unhover(aaa201)
         expect(path_101_201.classList.contains("inactive")).toBe(true)
       })
 
       it("hovering over a node should highlight all the missing nodes and not change the active nodes", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa202 = graph.getByTestId("aaa202")
         const aaa101 = graph.getByTestId("aaa101")
@@ -189,9 +204,9 @@ describe("Edge", () => {
         const path_aaa102_bool2 = graph.getByTestId("aaa102->bool2")
         const path_aaa201_bool2 = graph.getByTestId("aaa201->bool2")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa101)
+        await user.click(aaa101)
         expect(path_101_201.classList.contains("takeable")).toBe(true)
-        fireEvent.mouseEnter(aaa202)
+        await user.hover(aaa202)
         expect(path_101_201.classList.contains("takeable")).toBe(true)
         expect(path_bool2_aaa202.classList.contains("missing")).toBe(true)
         expect(path_aaa102_bool2.classList.contains("missing")).toBe(true)
@@ -199,46 +214,52 @@ describe("Edge", () => {
       })
 
       it("hovering over a node with missing edges then clicking it change the edge status to inactive", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa202 = graph.getByTestId("aaa202")
         const path_bool2_aaa202 = graph.getByTestId("bool2->aaa202")
-        fireEvent.mouseEnter(aaa202)
+        await user.hover(aaa202)
         expect(path_bool2_aaa202.classList.contains("missing")).toBe(true)
-        fireEvent.click(aaa202)
+        await user.click(aaa202)
         expect(path_bool2_aaa202.classList.contains("inactive")).toBe(true)
       })
     })
 
     describe("hovering over destination", () => {
       it("unselected source and (unselected and hovered destination) => edge should transition from 'inactive' to 'missing'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
 
         expect(path_101_201.classList.contains("inactive")).toBe(true)
-        fireEvent.mouseOver(aaa201)
+        await user.hover(aaa201)
         expect(path_101_201.classList.contains("missing")).toBe(true)
       })
       it("selected source and (unselected and hovered destination) => edge remains 'takeable'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa101 = graph.getByTestId("aaa101")
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa101)
+        await user.click(aaa101)
         expect(path_101_201.classList.contains("takeable")).toBe(true)
-        fireEvent.mouseOver(aaa201)
+        await user.hover(aaa201)
         expect(path_101_201.classList.contains("takeable")).toBe(true)
       })
       it("unselected source and (selected and hovered destination) => edge should transition to inactive 'missing'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa201)
+        await user.click(aaa201)
         expect(path_101_201.classList.contains("inactive")).toBe(true)
-        fireEvent.mouseOver(aaa201)
+        await user.unhover(aaa201) // needed to allow hover after click
+        await user.hover(aaa201)
         expect(path_101_201.classList.contains("missing")).toBe(true)
       })
       it("unselected source and (selected and hovered destination) => edge and all unmet and inactive prereqs should be 'missing'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa303 = graph.getByTestId("aaa303")
         const path_101_201 = graph.getByTestId("h101->aaa201")
@@ -246,14 +267,16 @@ describe("Edge", () => {
         const path_102_and = graph.getByTestId("aaa102->bool1")
         const path_and_303 = graph.getByTestId("bool1->aaa303")
 
-        fireEvent.click(aaa303) // selected node with missing prereqs
+        await user.click(aaa303) // selected node with missing prereqs
+        await user.unhover(aaa303)
 
         expect(path_101_201.classList.contains("inactive")).toBe(true)
         expect(path_201_and.classList.contains("inactive")).toBe(true)
         expect(path_102_and.classList.contains("inactive")).toBe(true)
         expect(path_and_303.classList.contains("inactive")).toBe(true)
 
-        fireEvent.mouseOver(aaa303)
+        await user.hover(aaa303)
+
         expect(path_101_201.classList.contains("missing")).toBe(true)
         expect(path_201_and.classList.contains("missing")).toBe(true)
         expect(path_102_and.classList.contains("missing")).toBe(true)
@@ -261,14 +284,15 @@ describe("Edge", () => {
       })
 
       it("selected source and selected destination, hovering over the destination will have the Edge remain 'active'", async () => {
+        const user = userEvent.setup()
         const graph = await TestGraph.build()
         const aaa101 = graph.getByTestId("aaa101")
         const aaa201 = graph.getByTestId("aaa201")
         const path_101_201 = graph.getByTestId("h101->aaa201")
-        fireEvent.click(aaa101)
-        fireEvent.click(aaa201)
+        await user.click(aaa101)
+        await user.click(aaa201)
         expect(path_101_201.classList.contains("active")).toBe(true)
-        fireEvent.mouseOver(aaa201)
+        await user.hover(aaa201)
         expect(path_101_201.classList.contains("active")).toBe(true)
       })
     })
@@ -281,6 +305,7 @@ describe("Edge", () => {
       })
     })
     it("should restore edge status", async () => {
+      const user = userEvent.setup()
       const graph = await TestGraph.build()
       const aaa102 = graph.getByTestId("aaa102")
       const aaa201 = graph.getByTestId("aaa201")
@@ -290,8 +315,8 @@ describe("Edge", () => {
       const aaa201_bool2 = graph.getByTestId("aaa201->bool2")
       const aaa201_bool1 = graph.getByTestId("aaa201->bool1")
       const bool1_aaa303 = graph.getByTestId("bool1->aaa303")
-      fireEvent.click(aaa102)
-      fireEvent.click(aaa201)
+      await user.click(aaa102)
+      await user.click(aaa201)
       expect(aaa102_bool2.classList.contains("active")).toBe(true)
       expect(aaa102_bool1.classList.contains("active")).toBe(true)
       expect(bool2_aaa202.classList.contains("takeable")).toBe(true)